+++ src/inject/detector.ts
@@ -1,11 +1,11 @@
-import {parse, getSRGBLightness} from '../utils/color';
+import {getSRGBLightness, parseColorWithCache} from '../utils/color';
 
 function hasBuiltInDarkTheme() {
     const drStyles = document.querySelectorAll('.darkreader') as NodeListOf<HTMLStyleElement & {disabled: boolean}>;
     drStyles.forEach((style) => style.disabled = true);
 
-    const rootColor = parse(getComputedStyle(document.documentElement).backgroundColor);
-    const bodyColor = document.body ? parse(getComputedStyle(document.body).backgroundColor) : {r: 0, g: 0, b: 0, a: 0};
+    const rootColor = parseColorWithCache(getComputedStyle(document.documentElement).backgroundColor);
+    const bodyColor = document.body ? parseColorWithCache(getComputedStyle(document.body).backgroundColor) : {r: 0, g: 0, b: 0, a: 0};
     const rootLightness = (1 - rootColor.a) + rootColor.a * getSRGBLightness(rootColor.r, rootColor.g, rootColor.b);
     const finalLightness = (1 - bodyColor.a) * rootLightness + bodyColor.a * getSRGBLightness(bodyColor.r, bodyColor.g, bodyColor.b);
     const darkThemeDetected = finalLightness < 0.5;