+++ src/inject/dynamic-theme/modify-css.ts
@@ -1,5 +1,4 @@
-import type {RGBA} from '../../utils/color';
-import {lowerCalcExpression, parse, rgbToHSL, hslToString} from '../../utils/color';
+import {parseColorWithCache, rgbToHSL, hslToString} from '../../utils/color';
 import {clamp} from '../../utils/math';
 import {getMatches} from '../../utils/text';
 import {getAbsoluteURL} from '../../utils/url';
@@ -140,7 +139,7 @@ export function getSelectionColor(theme: Theme) {
         backgroundColorSelection = modifyBackgroundColor({r: 0, g: 96, b: 212}, {...theme, grayscale: 0});
         foregroundColorSelection = modifyForegroundColor({r: 255, g: 255, b: 255}, {...theme, grayscale: 0});
     } else {
-        const rgb = parse(theme.selectionColor);
+        const rgb = parseColorWithCache(theme.selectionColor);
         const hsl = rgbToHSL(rgb);
         backgroundColorSelection = theme.selectionColor;
         if (hsl.l < 0.5) {
@@ -182,7 +181,7 @@ function getModifiedScrollbarStyle(theme: Theme) {
         colorThumbActive = modifyBackgroundColor({r: 96, g: 96, b: 96}, theme);
         colorCorner = modifyBackgroundColor({r: 255, g: 255, b: 255}, theme);
     } else {
-        const rgb = parse(theme.scrollbarColor);
+        const rgb = parseColorWithCache(theme.scrollbarColor);
         const hsl = rgbToHSL(rgb);
         const isLight = hsl.l > 0.5;
         const lighten = (lighter: number) => ({...hsl, l: clamp(hsl.l + lighter, 0, 1)});
@@ -237,31 +236,6 @@ const unparsableColors = new Set([
     'none',
     'unset',
 ]);
-
-const colorParseCache = new Map<string, RGBA>();
-
-export function parseColorWithCache($color: string) {
-    $color = $color.trim();
-    if (colorParseCache.has($color)) {
-        return colorParseCache.get($color);
-    }
-    // We cannot _really_ parse any color which has the calc() expression
-    // So we try our best-efforts to remove those and then parse the value.
-    if ($color.includes('calc(')) {
-        $color = lowerCalcExpression($color);
-    }
-    const color = parse($color);
-    colorParseCache.set($color, color);
-    return color;
-}
-
-export function tryParseColor($color: string) {
-    try {
-        return parseColorWithCache($color);
-    } catch (err) {
-        return null;
-    }
-}
 
 function getColorModifier(prop: string, value: string): string | CSSValueModifier {
     if (unparsableColors.has(value.toLowerCase())) {
@@ -348,20 +322,20 @@ export function getBgImageModifier(
             const parts = getMatches(partsRegex, match, 1).map((part) => {
                 part = part.trim();
 
-                let rgb = tryParseColor(part);
+                let rgb = parseColorWithCache(part);
                 if (rgb) {
                     return (filter: FilterConfig) => modifyGradientColor(rgb, filter);
                 }
 
                 const space = part.lastIndexOf(' ');
-                rgb = tryParseColor(part.substring(0, space));
+                rgb = parseColorWithCache(part.substring(0, space));
                 if (rgb) {
                     return (filter: FilterConfig) => `${modifyGradientColor(rgb, filter)} ${part.substring(space + 1)}`;
                 }
 
                 const colorStopMatch = part.match(colorStopRegex);
                 if (colorStopMatch) {
-                    rgb = tryParseColor(colorStopMatch[3]);
+                    rgb = parseColorWithCache(colorStopMatch[3]);
                     if (rgb) {
                         return (filter: FilterConfig) => `${colorStopMatch[1]}(${colorStopMatch[2] ? `${colorStopMatch[2]}, ` : ''}${modifyGradientColor(rgb, filter)})`;
                     }
@@ -503,7 +477,7 @@ export function getShadowModifierWithInfo(value: string): CSSValueModifierWithIn
             const matchIndex = value.indexOf(match, index);
             const matchEnd = matchIndex + match.length;
             index = matchEnd;
-            const rgb = tryParseColor(match);
+            const rgb = parseColorWithCache(match);
             if (!rgb) {
                 notParsed++;
                 return () => value.substring(prefixIndex, matchEnd);
@@ -559,7 +533,6 @@ function getVariableDependantModifier(
 }
 
 export function cleanModificationCache() {
-    colorParseCache.clear();
     clearColorModificationCache();
     imageDetailsCache.clear();
     cleanImageProcessingCache();