+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Correctly open static theme editor on Mobile.
 - Migrate automation settings to it's own object. WARNING: this can lead to data loss of currently used automation settings due to browser behavior.
 - Correctly handle empty URL's in `background-image` property.
+- Make parsing colors use cache.
 
 ## 4.9.52 (June 28, 2022)